fix: move client_ctx under core
