+++ crates/ibc-testkit/src/testapp/ibc/clients/mod.rs
@@ -1,2 +1,1 @@
-pub mod context;
 pub mod types;
