+++ crates/ibc-testkit/src/testapp/ibc/core/mod.rs
@@ -1,4 +1,5 @@
+pub mod client_ctx;
 pub mod configs;
-pub mod context;
+pub mod core_ctx;
 pub mod router;
 pub mod types;
